pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "lng-market-topology"
version = "0.1.0"
description = "Constraint-based inference for global LNG market counterparty behavior"
requires-python = ">=3.10"
dependencies = [
    "networkx>=3.0",
    "pyyaml>=6.0",
    "numpy>=1.24",
    "pandas>=2.0",
    "matplotlib>=3.7",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
